Fix for `segment_list` response outputting 'An unexpected problem occurred ...' error message.
This is a v2 API call that doesn't have the `result_code` or `result_message` keys in the response, so we assume something invalid came back in the API response.

We really just needed to whitelist this API method (like we do for other methods that don't include that extra information).

DEVJIRA-2064